Add files via upload
This is our true final code because the previous code had basically nothing to do with the whole FIFA problem and was a basic scheduling example. The formulation I have basically just ignores the teams and just assigns the matches directly and it still takes care of the 3 days of rest and does it according to the FIFA schedule too but I can only make it work for 1 group at a time so basically we'll just have to manually do it per group, which is fine. There's really no other way to do it either because for any scheduling problem its nearly impossible to have it work 100% purely through code without some manual tweaking